Fix API bugs: Convert all routes to unified createApiHandler pattern
- Fixed template creation route mismatch (POST /template now works)
- Fixed parameter destructuring bugs in dynamic routes ([templateId], [sessionId])
- Converted all remaining routes to use createApiHandler/createValidatedApiHandler
- Removed duplicate /template/create route
- Fixed inconsistent error handling and response formats
- Added environment variable examples (.env.example)
- Ensured consistent API response structure across all endpoints

Routes converted:
- /api/template/[templateId]/* (GET, PUT, DELETE, complete, favorite)
- /api/session/[sessionId]/* (GET, PUT, DELETE)
- /api/session/active/recover (POST)
- /api/session/scheduled (GET)
- /api/analytics (GET)
- /api/dashboard (GET)
- /api/search (GET)
- /api/user/* (achievements, dashboard-config, stats)

All API routes now use unified authentication, validation, and error handling.
