+++ .env.example
@@ -1,0 +1,11 @@
+# Database Configuration
+DATABASE_URL="postgresql://username:password@localhost:5432/eternal_fitness"
+DIRECT_URL="postgresql://username:password@localhost:5432/eternal_fitness"
+
+# Clerk Authentication
+NEXT_PUBLIC_CLERK_PUBLISHABLE_KEY="pk_test_your_clerk_publishable_key_here"
+CLERK_SECRET_KEY="sk_test_your_clerk_secret_key_here"
+SIGNING_SECRET="whsec_your_clerk_webhook_signing_secret_here"
+
+# Optional: Node Environment
+NODE_ENV="development"
