+++ src/app/api/session/active/recover/route.ts
@@ -1,19 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { auth } from '@clerk/nextjs/server';
 import prisma from '@/lib/prisma';
-
-// Response helpers
-const successResponse = (data: any, status = 200) => {
-  return NextResponse.json(data, { status });
-};
-
-const errorResponse = (message: string, status = 500, details?: any) => {
-  console.error(`API Error (${status}):`, message, details ? JSON.stringify(details) : '');
-  return NextResponse.json(
-    { error: { message, ...(details && { details }) } },
-    { status }
-  );
-};
+import { createValidatedApiHandler } from '@/lib/api-utils';
 import { 
   ActiveWorkoutSessionData, 
   WorkoutTemplateData 
@@ -33,21 +21,9 @@ const recoverSessionSchema = z.object({
 // POST: Recover or reset active workout session
 // ============================================================================
 
-export async function POST(request: NextRequest) {
-  try {
-    const { userId } = await auth();
-    if (!userId) {
-      return errorResponse('Unauthorized', 401);
-    }
-
-    const body = await request.json();
-    const validationResult = recoverSessionSchema.safeParse(body);
-
-    if (!validationResult.success) {
-      return errorResponse('Invalid recovery data', 400, validationResult.error.errors);
-    }
-
-    const { templateId, forceRecover } = validationResult.data;
+export const POST = createValidatedApiHandler(
+  recoverSessionSchema,
+  async (userId, { templateId, forceRecover }) => {
 
     // Get current active session
     const userStats = await prisma.userStats.findUnique({
@@ -60,7 +36,7 @@ export async function POST(request: NextRequest) {
     });
 
     if (!userStats?.activeWorkoutId || !userStats.activeWorkoutData) {
-      return errorResponse('No active workout session found to recover', 404);
+      throw new Error('No active workout session found to recover');
     }
 
     const currentSessionData = userStats.activeWorkoutData as unknown as ActiveWorkoutSessionData;
@@ -93,13 +69,15 @@ export async function POST(request: NextRequest) {
       issues.push('Template no longer exists or is not accessible');
     }
 
-    // If there are issues and force recover is not enabled, return the issues
+    // If there are issues and force recover is not enabled, throw error with details
     if (issues.length > 0 && !forceRecover) {
-      return errorResponse('Session data integrity issues found', 400, {
+      const error = new Error('Session data integrity issues found');
+      (error as any).details = {
         issues,
         canRecover: !!template,
         suggestion: 'Use forceRecover=true to attempt recovery or clear the session',
-      });
+      };
+      throw error;
     }
 
     // Attempt recovery if template exists
@@ -133,12 +111,12 @@ export async function POST(request: NextRequest) {
         },
       });
 
-      return successResponse({ 
+      return {
         activeSession: recoveredSessionData,
         recovered: true,
         issues: issues.length > 0 ? issues : undefined,
         message: 'Active workout session recovered successfully'
-      });
+      };
     }
 
     // If we can't recover, clear the session
@@ -151,15 +129,11 @@ export async function POST(request: NextRequest) {
       },
     });
 
-    return successResponse({ 
+    return {
       activeSession: null,
       recovered: false,
       issues,
       message: 'Active workout session cleared due to unrecoverable issues'
-    });
-
-  } catch (error) {
-    console.error('Error recovering active session:', error);
-    return errorResponse('Failed to recover active session', 500);
+    };
   }
-}
+);
