+++ src/app/api/session/scheduled/route.ts
@@ -1,31 +1,10 @@
 import { NextResponse } from 'next/server';
 import { auth } from '@clerk/nextjs/server';
 import prisma from '@/lib/prisma';
-
-// --- Standard Response Helpers ---
-const successResponse = (data: any, status = 200) => {
-  return NextResponse.json({ data }, { status });
-};
-
-const errorResponse = (message: string, status = 500, details?: any) => {
-  console.error(
-    `API Error (${status}) [session/scheduled]:`,
-    message,
-    details ? JSON.stringify(details) : '',
-  );
-  return NextResponse.json(
-    { error: { message, ...(details && { details }) } },
-    { status },
-  );
-};
+import { createApiHandler } from '@/lib/api-utils';
 
 // GET function to fetch ONLY scheduled sessions
-export async function GET() {
-  try {
-    const { userId } = await auth();
-    if (!userId) {
-      return errorResponse('Unauthorized', 401);
-    }
+export const GET = createApiHandler(async (userId) => {
 
     const scheduledSessions = await prisma.workoutSession.findMany({
       where: {
@@ -54,13 +33,5 @@ export async function GET() {
       },
     });
 
-    return successResponse(scheduledSessions);
-  } catch (error) {
-    console.error('Error in GET /api/session/scheduled:', error); // Keep specific logging
-    return errorResponse(
-      'Internal Server Error',
-      500,
-      error instanceof Error ? error.message : String(error),
-    );
-  }
-}
+    return scheduledSessions;
+});
